+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
  - Runic is now silent by default. Use `--verbose` to enable the verbose file progress
    printing from previous releases ([#121]).
+### Fixed
+ - Improved the command line argument parsing to handle more edge cases related to input
+   files and directories such as e.g. empty directories ([#123]).
 
 ## [v1.2.0] - 2024-12-09
 ### Added